+++ src/ast/index.ts
@@ -1,4 +1,5 @@
 export * from "./domain";
 export * from "./scope";
 export * from "./selection";
+export * from "./siblings";
 export * from "./transformation";