+++ src/ast/siblings.ts
@@ -1,0 +1,35 @@
+import { NodePath } from "@babel/traverse";
+import * as t from "@babel/types";
+
+export {
+  hasSiblingStatement,
+  getPreviousSiblingStatements,
+  getNextSiblingStatements
+};
+
+function hasSiblingStatement(path: NodePath): boolean {
+  const allSiblingStatements = [
+    ...getPreviousSiblingStatements(path),
+    ...getNextSiblingStatements(path)
+  ];
+
+  return allSiblingStatements.length > 0;
+}
+
+function getPreviousSiblingStatements(path: NodePath): t.Statement[] {
+  return path
+    .getAllPrevSiblings()
+    .map(({ node }) => node)
+    .filter(isStatement);
+}
+
+function getNextSiblingStatements(path: NodePath): t.Statement[] {
+  return path
+    .getAllNextSiblings()
+    .map(({ node }) => node)
+    .filter(isStatement);
+}
+
+function isStatement(node: t.Node): node is t.Statement {
+  return t.isStatement(node);
+}